requirements.txt
atomicwrites==1.3.0
attrs==19.1.0
backcall==0.1.0
decorator==4.4.0
Django==2.1.7
djangorestframework==3.9.2
ipdb==0.12
ipython==7.3.0
ipython-genutils==0.2.0
jedi==0.13.3
more-itertools==6.0.0
parso==0.3.4
pexpect==4.6.0
pickleshare==0.7.5
pkg-resources==0.0.0
pluggy==0.9.0
prompt-toolkit==2.0.9
psycopg2==2.7.7
psycopg2-binary==2.7.7
ptyprocess==0.6.0
py==1.8.0
Pygments==2.3.1
pytest==4.3.1
pytest-django==3.4.8
pytz==2018.9
six==1.12.0
traitlets==4.3.2
wcwidth==0.1.7
